+++ functions.php
@@ -83,7 +83,7 @@ function tenores_replace_my_account_menu_text($title, $item, $args, $depth)
 {
     if (isset($args->theme_location) && $args->theme_location === 'primary' && is_user_logged_in()) {
         $my_account_url = home_url('/minha-conta');
-        
+
         if ($item->url === $my_account_url || $item->url === trailingslashit($my_account_url)) {
             return 'Minha Conta';
         }
@@ -93,3 +93,11 @@ function tenores_replace_my_account_menu_text($title, $item, $args, $depth)
 }
 
 add_filter('nav_menu_item_title', 'tenores_replace_my_account_menu_text', 10, 4);
+
+function fa_remove_password_strength()
+{
+    wp_dequeue_script('wc-password-strength-meter');
+    wp_deregister_script('wc-password-strength-meter');
+}
+
+add_action('wp_enqueue_scripts', 'fa_remove_password_strength', 99999);